Add optional `virtual` to `DigitalIn/Out/InOut` methods
This commit introduces configuration options to make certain methods of the `DigitalIn`, `DigitalOut`, and `DigitalInOut` classes virtual. This is useful when more advanced features of polymorphism are required. This is not enabled by default to optimize these classes for execution speed and code size.

Enabling this option will slightly increase latency of calling the virtual methods and will increase the ROM/RAM footprint of the class.

The configuration can be changed by overriding the options: `drivers.virtualize-digitalin` (and so on, for each class involved in the commit).